Break out teams block component
