+++ components/Game/index.ts
@@ -1,7 +1,9 @@
 import DivisionStandings from './Preview/DivisionStandings';
+import TeamsBlock from './Preview/TeamsBlock';
 import TeamStats from './Preview/TeamStats';
 
 export {
   DivisionStandings,
+  TeamsBlock,
   TeamStats
 };
